Fixed more comments, and added a cutoff switch
